mcp-lambda/requirements.txt
awslabs.mcp-lambda-handler>=0.1.14
boto3>=1.34.0
